requirements.txt
# Core dependencies - Agent Framework (REQUIRES --pre flag during preview)
agent-framework-azure-ai

# Azure SDKs
azure-ai-formrecognizer>=3.3.0
azure-identity>=1.15.0
azure-ai-inference>=1.0.0b1

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0
pypdf>=3.17.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web server for MCP
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Async support
aiohttp>=3.9.0
